Fix getting last commit from the repo
We have excluded merge commits for the log in df192c2 to avoid
using the 'artificial' merge commit produced during the build
process but sometimes a merge commit can be a valid last commit
in the repo we are trying to build so we need to exclude commits
based on the author instead. Our merge commits are commited by
'Jenkins' so it should be safe to exclude all commits based on
this.